Split runtime event runner from hook adapter
