+++ backend/threads/chat_adapters/runtime_event_hook.py
@@ -1,17 +1,8 @@
 from __future__ import annotations
 
 import asyncio
-from collections.abc import Callable, Coroutine, Iterable
+from collections.abc import Callable, Coroutine
 from typing import Any
-
-
-async def run_planned_runtime_event[EventT, ActionT, ResultT](
-    event: EventT,
-    planner: Callable[[EventT], Iterable[ActionT]],
-    dispatch_actions: Callable[[list[ActionT]], Coroutine[Any, Any, ResultT]],
-) -> ResultT:
-    actions = list(planner(event))
-    return await dispatch_actions(actions)
 
 
 def make_sync_runtime_event_hook[**P](async_fn: Callable[P, Coroutine[Any, Any, Any]]) -> Callable[P, None]:
